Copy the DLLs from narupa-protocol
This commit copies the DLLs produced in PR 34 for narupa-protocol where
gRPC and protobuff are updated. The update of gRPC should be especially
impactful as it moves from the gRPC-core-C library developed by Google
to the gRPC for .NET library developed my Microsoft.

There may be too manu DLLs copied from the narupa-protocol build.
Indeed, many of these files are new while unrelated to the changes in
narupa-protocol. There must have been a selection process for the DLLs
to copy, but I am not aware of it.