Move the chemical component dictionary to a text asset
The CIF importer has a library of bonds for the residues of the PDB.
This is initially read from the component library provided by the PDB,
then serialised, compressed, and encoded as base64.

The class handling the library is `ChemicalComponentDictionary` which,
before this commit, had the serialised library in a serialised field and
was loading it from an asset that was a serialised instance of the
class.

Unity does not like this setup and the tests were failing with this
error repeated:

> Recursive Serialization is not supported. You can't dereference a PPtr while loading. (Constructors of C# classes may not load objects either. See stacktrace.)

Here, we move the base64 encoded library from a
`ChemicalComponentDictionary` asset to a `TextAsset` that is loaded from
the `ChemicalComponentDictionary` class. This avoid the recursive
serialisation issue.