Move the controller assets in a dedicated directory
This commit have `SteamVrControllerDefinition` look for the controller
assets in a `Controllers` directory in any `Resources` folder rather
than directly in a `Resources` folder. This limits what unity tries to
load when looking for the controller. Otherwise, Unity tries to load all
the assets before keeping only the controllers.

In Narupa iMD, this means the controller assets need to move from
`Assets/NarupaIMD/Resources` to
`Assets/NarupaIMD/Resources/Controllers`.