Run the tests on the plugin in CI
The tests in the plugin already run for the Narupa iMD repository,
but running them for the plugin itself make sure that the plugin can run
out of the context of a specific application.

To run, the tests require the `host-project` branch. The tested branch
is checked out *within* the content of the `host-project` branch to
mimic a project using the plugin and copying it in
`Assets/Plugins/Narupa`.