Remove the SteamVrControllerDefinitionTests
The tests that make sure the host project implements the assets for a set of controllers depends on the files
in the host project. This mean they cannot pass when running in the plugin itself. Ultimately, it is the
responsibility of the host project to choose what controllers to implement.

Narupa IMD implements the deleted tests in https://github.com/IRL2/narupa-imd/pull/12